examples/avatar/simli_avatar/requirements.txt
fastapi
uvicorn
opencv-python
httpx
python-dotenv
numpy
sounddevice
aiortc[video]
av
websockets
wave

simli-ai[livekit]
livekit>=0.9.0
livekit-agents>=0.5.0

opencv-python-headless